fix clippy issue
