+++ ink/logics/traits/participant_filter/mod.rs
@@ -1,2 +1,16 @@
 pub mod filter_latest_winners;
-pub mod participant_filter;
+
+use openbrush::contracts::access_control::AccessControlError;
+
+#[derive(Debug, Eq, PartialEq, scale::Encode, scale::Decode)]
+#[cfg_attr(feature = "std", derive(scale_info::TypeInfo))]
+pub enum ParticipantFilterError {
+    AccessControlError(AccessControlError),
+}
+
+/// convertor from AccessControlError to ParticipantFilterError
+impl From<AccessControlError> for ParticipantFilterError {
+    fn from(error: AccessControlError) -> Self {
+        ParticipantFilterError::AccessControlError(error)
+    }
+}
